Adds Timer and Message classes, FadeAfter...
behaviour to fade something after a delay.

The plan is to work the timer into all the classes that need it.

Timers will be able (eventually) to check Game.paused in order to
update, and will have an option to tick even when paused.

The Message class stores a string, and can be drawn on the screen. Has
support for text colors, using escape codes inside of the string (ex.
"\\r" or literally (as it would appear normally, \r) to change the text
color to red.)
Multiple text codes are supported, and support for more will be added in
the future.