Updates
Reworks message to be serializable, and has an exposed style parameter
which can be changed to change the appearance of the message when drawn.
Reworks GUIMessage to use the Message class.

Adds mouseRay accessor into CameraF, and adds mouseCollider accessor.

Adds ImageChopper EditorWindow class. This is a wizard which chops up an
image into a bunch of smaller images, dumping them into a resources
folder with the original image name + a number after, to make them easy
to load.