Holy mother of fuck lots of changes
Makes items have an ID property, and ItemDatabase searches by ID first,
then name.

Renames ItemDatabase to ItemEditorWindow so it loads when the project is
re-opened.

Adds speed field to FancyBar

Optimizes foreach loops in Table to loop through the KeyValue pairs
instead of just Keys

Adds flags class blueprint - acts like a table for booleans. attempting
to get it to save/load to player prefs in a highly compressed format.

Planning to do a massive rework of things- lots of stuff will be cut,
and lots of stuff will be adjusted in order to be better organized and
make more sense.

When that happens, many projects that rely on legacy behaviour of this
library may not be supported.

Some ideas were to mask classes that unity has with our own
implementation (EX instead of RandomF, the class would be called Random,
and simply exist on-top of unity's Random class, providing the same
functions, with a number of extra functions, and behaviour adjusted to
match what we want. Same with ScreenF, InputF, etc. Extension method
classes would be renamed to 'VectorUtilities' instead of VectorF, etc.)